Move module `parquet_rs` to `parquet_rs_old`.
